frontend/requirements.txt
# Core bot framework
aiogram==3.13.1

# Environment variables
python-dotenv==1.0.1

# HTTP requests (for backend API calls)
requests==2.32.3
httpx==0.27.2

# Async support (aiogram dependencies)
aiohttp==3.10.5
aiofiles==23.2.1
magic-filter==1.0.12

# Data validation
pydantic==2.9.2

# URL handling
yarl==1.13.1
idna==3.10

# Additional aiogram dependencies
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
frozenlist==1.4.1
multidict==6.1.0
typing-extensions==4.12.2
